Move YouTubeChannel component into features to allow wrapping with context provider
